requirements.txt
# LLM 相关
openai>=1.0.0
tiktoken>=0.9.0  # Token 计数

# 工具库
pydantic>=2.0.0
pyyaml>=6.0.0
dataclasses-json>=0.6.0

# 可选：本地模型支持（未来）
# vllm>=0.6.0  # 用于本地模型推理
# ollama>=0.1.0  # Ollama 客户端

# 开发工具（可选）
# pytest>=7.0.0
# black>=23.0.0
# mypy>=1.0.0

